add region mapping for older s2 datasets
